Completed first version of EA emails script. Changed scrape wait to 1 second - this is not a hugely limiting factor

+++ R_analysis/ea_community_emails.R
@@ -2,7 +2,7 @@ dir.create(email_tables, showWarnings = FALSE)
 
 min_cause_area_sum <- 10 #make this cause area specific?
 ea_cause_areas <- c('Biorisk', 'China', 'Climate change', 
-                    'Factory farming', 'Nuclear') 
+                    'Factory farming', 'Nuclear', 'AI', 'Mental health', 'Global poverty') 
 #These should bepart of the cause area info and key word search
 #The lines above are a quick fix - without them we have lots of DfID jobs
 
@@ -10,20 +10,33 @@ all_advert_data <-adverts_csv_name %>%
   read_csv
 
 needed_advert_info <- all_advert_data %>%
-  select(job_ref, `Number of posts`)
+  select(job_ref, `Number of posts`) 
+
+
+previously_identified_adverts <- email_tables %>%
+  list.files() %>%
+  map(~read_csv(file.path(email_tables, .))) %>%
+  reduce(bind_rows) %>%
+  pull(link) 
 
 emailable_key_words_results <- key_words_results_file %>%
   read_csv %>%
   left_join(needed_advert_info) %>%
+  mutate(`To share with EAs? Y/N/?` = '',
+         Notes = '') %>%
   filter(cause_area_sum >= min_cause_area_sum,
-         closing_date >= today(),
-         `Cause area` %in% ea_cause_areas) %>%
-  select(`Cause area`, job_department,
-         link, grade, location, 
-         `Number of posts`, date_downloaded,
-         closing_date, job_title, approach) %>%
-  mutate(`Toby’s subjective and probably terrible rating of how impactful this job is within this cause area: please do not take this seriously he is not very smart and is a generally terrible person` = '')
+         (closing_date - today()) > 4,
+         `Cause area` %in% ea_cause_areas,
+         approach == 'External',
+         !(link %in%  previously_identified_adverts)) %>%
+  select(`Cause area`,
+         link,
+         job_department,
+         `To share with EAs? Y/N/?`,
+         Notes)
 
+#check if there are any new ones?
 my_file_name <- file.path(email_tables, paste(today(),'.csv', sep = ''))
-
-write_csv(emailable_key_words_results, my_file_name)
+if(!file.exists(my_file_name)){
+  write_csv(emailable_key_words_results, my_file_name)
+} else(print('File already exists delete/rename and rerun!'))
